Add files via upload
Data analysis, and text analysis, have come into the crosshairs of every business interested in ongoing success in modern markets. Quite literally, the ability to analyze and quantify customer reviews and other customer info has become the coin of the realm for businesses looking to make a name for themselves.